Line Rider Clone for PS Vita
Using the University's own framework and Box2D, a basic Line Rider Clone
was created. It uses Bezier Curve mathematics to calculate the curve the
player has drawn and dynamically adds physics bodies to it. The player
is able to draw curves, straight lines and erase parts of each to create
a physics 'playground'. They are then able to send controllable or
free-falling physics bodies across their designs just as in Line Rider.

Note: The application will not run without a) a PS Vita development kit
and b) the Abertay Framework.