Side-scrolling Beat 'em Up for PS Vita
Coursework completed for Semester 1 of year 2. This coursework required
the use of the Vita's controls with the use of sprites and collision
detection. This was taken further with sprite animation. Sprites have
been ripped from a TMNT SNES game. This application is not intended for
sale or public consumption of any kind under any circumstances.