Spelunky Level Generation Tool
Coursework completed for Semester 1 of Year 3 for the Games Tool
Programming module. This tool, written in C, uses an algorithm based off
of the description for Derek Yu's Spelunky level generator. It
incorporates only the initial section of the generator where it takes a
start point in a grid and finds a path to the exit. The included report
details the tool and its uses in far greater detail.